refactor: remove lots of folders and unify export
