+++ packages/graphql-ui/src/components/Content.tsx
