+++ packages/graphql-ui/src/components/SchemaEditor.tsx
@@ -24,41 +24,22 @@ import { GraphQLSchema } from "graphql";
 import { Button, Checkbox } from "@neo4j-ndl/react";
 import * as neo4j from "neo4j-driver";
 import { EditorFromTextArea } from "codemirror";
-import { CodeMirror } from "../../utils/utils";
-import * as AuthContext from "../../contexts/auth";
+import { CodeMirror } from "../utils/utils";
+import * as AuthContext from "../contexts/auth";
 import {
+    DEFAULT_OPTIONS,
+    DEFAULT_TYPE_DEFS,
     LOCAL_STATE_CHECK_CONSTRAINT,
     LOCAL_STATE_CREATE_CONSTRAINT,
     LOCAL_STATE_DEBUG,
     LOCAL_STATE_TYPE_DEFS,
     SCHEMA_EDITOR_BUILD_BUTTON,
     SCHEMA_EDITOR_INPUT,
-} from "../../constants";
-import { formatCode, ParserOptions } from "../editor/utils";
+} from "../constants";
+import { formatCode, ParserOptions } from "./editor/utils";
 import { Col, ColsWrapper, Row, RowsWrapper } from "react-grid-resizable";
-import { JSONEditor } from "../editor/JSONEditor";
-import { Extension, FileName } from "../editor/Filename";
-
-const DEFAULT_TYPE_DEFS = `
-    # Write your own type definition in the editor here or 
-    # generate it automatically from the current Neo4j database (introspection)
-
-    # Example type definition:
-    type Movie {
-        title: String!
-    }
-    
-`;
-
-const DEFAULT_OPTIONS = JSON.stringify(
-    {
-        config: {
-            enableRegex: true,
-        },
-    },
-    null,
-    2
-);
+import { JSONEditor } from "./editor/JSONEditor";
+import { Extension, FileName } from "./editor/Filename";
 
 export interface Props {
     onChange: (s: GraphQLSchema) => void;
